+++ pkg/bootstrap/versions/v2_0_1/cluster_upgrade_list.go
@@ -27,6 +27,8 @@ var needMigrateMoPubs = false
 var clusterUpgEntries = []versions.UpgradeEntry{
 	upg_mo_table_stats,
 	upg_mo_pubs_add_account_id_column,
+	upg_mo_cdc_task,
+	upg_mo_cdc_watermark,
 }
 
 var upg_mo_table_stats = versions.UpgradeEntry{
@@ -51,6 +53,37 @@ var upg_mo_pubs_add_account_id_column = versions.UpgradeEntry{
 		}
 
 		needMigrateMoPubs = !colInfo.IsExits
+		return colInfo.IsExits, nil
+	},
+}
+
+var upg_mo_cdc_task = versions.UpgradeEntry{
+	Schema:    catalog.MO_CATALOG,
+	TableName: catalog.MO_CDC_TASK,
+	UpgType:   versions.CHANGE_COLUMN,
+	UpgSql:    fmt.Sprintf("alter table %s.%s change reserved1 err_msg varchar(256)", catalog.MO_CATALOG, catalog.MO_CDC_TASK),
+	CheckFunc: func(txn executor.TxnExecutor, accountId uint32) (bool, error) {
+		colInfo, err := versions.CheckTableColumn(txn, accountId, catalog.MO_CATALOG, catalog.MO_CDC_TASK, "err_msg")
+		if err != nil {
+			return false, err
+		}
+		return colInfo.IsExits, nil
+	},
+}
+
+var upg_mo_cdc_watermark = versions.UpgradeEntry{
+	Schema:    catalog.MO_CATALOG,
+	TableName: catalog.MO_CDC_WATERMARK,
+	UpgType:   versions.ADD_COLUMN,
+	UpgSql: fmt.Sprintf("alter table %s.%s "+
+		"add column db_name varchar(256) after table_id, "+
+		"add column table_name varchar(256) after db_name, "+
+		"add column err_msg varchar(256) after watermark", catalog.MO_CATALOG, catalog.MO_CDC_WATERMARK),
+	CheckFunc: func(txn executor.TxnExecutor, accountId uint32) (bool, error) {
+		colInfo, err := versions.CheckTableColumn(txn, accountId, catalog.MO_CATALOG, catalog.MO_CDC_WATERMARK, "db_name")
+		if err != nil {
+			return false, err
+		}
 		return colInfo.IsExits, nil
 	},
 }
